fix(dynamo-run): Text input doesn't need a name
For the `echo` and `pystr` engines we previously required the user to
pass `--model-name <x>` so we would have a name for the model. If the
input is HTTP we do need this to match on the users' JSON request.

If the input is Text we don't need a name. So if the input is text and
we don't already have a name for the model, give it one.