fix(vllm,sglang): Let the engine enforce max tokens
Previously several parts of the stack ensured max tokens (for this
single request) was set.

Now only text input sets it (to 8k). Everything else leaves as is, potentially blank.

The engines themselves have very small defaults, 16 for vllm and 128 for
sglang.

Also fix dynamo-run CUDA startup message to only print if we're using an
engine that would benefit from it (mistralrs, llamacpp).