feat: Python decorator dynamo_static_worker without etcd
Introduce `@dynamo_static_worker` as an alternative to `@dynamo_worker`.

Static workers (`dynamo_static_worker`) have a predictable
name and etcd is not required. Ee don't do any runtime discovery.

This contrasts with the dynamic `dynamo_worker` endpoints we have now,
which get a unique random name (based on namespace/component/endpoint),
and are discovered by ingress via etcd.

Change the hello_world example to use dynamo_static_worker so that it is
exercised and demonstrated somewhere.

For NIM.