chore: Split PushRouter from Client
In a distributed system we don't know if the remote workers need
pre-processing done ingress-side or not. Previously Client required us
to decide this before discovering the remote endpoints, which was fine
because pre-processing was worker-side.

As part of moving pre-processing back to ingress-side we need to split
this into two steps:
- Client discovers the endpoints, and (later PR) will fetch their Model
  Deployment Card.
- PushRouter will use the Model Depoyment Card to decide if they need
  pre-processing or not, which affects the types of the generic
  parameters.

Part of #743