chore: Prevent duplicate components with different models.
Each namespace is for a single pipeline, so a component must be
model-unique. The means we can have several components with the same
name running the same model (data parallel), their traffic will be
routed according to `--router-mode`, but we cannot have several
components with the same name running different models.

Add a check to prevent that happening.