fix(dynamo-run): Use a random endpoint for internal comms
To talk to the vllm/sglang/trtllm engine we previously hardcoded an
endpoint. The user never sees it so it doesn't matter which one.

However if you try to run _two_ instances of Dynamo on one machine they
will conflict.

Use a UUID as the component name to resolve that.

Part of the solution for:
https://github.com/ai-dynamo/dynamo/issues/1073