perf(runtime): Use all available parallelism
Performance work showed a drop-off in performance scaling when
concurrency exceeds 16. Which is the number of threads we were giving
tokio. Now we use all the cores on the machine which is the tokio default
(and also what Go does).

Also ensure in most cases we only use a single tokio runtime. I still contend
the secondary runtime is an unnecessary complication, but this way we preserve
optionality (in case we choose to pin the secondary runtime to a single
core), but also simplify the performance characteristics.