feat: Python frontend / ingress node
Before: `dynamo-run in=http out=dyn [flags]`
After: `python -m dynamo.ingress [flags]`

No need to build or install the dynamo-run Rust binary. This will have the
same performance as dynamo-run, it uses the Rust library.

Note we want to use `dynamo.frontend` but the older examples still use
that namespace. As soon as they are removed we can rename this.