chore: Remove service_name from ModelDeploymentCard
We had two names in the ModelDeploymentCard: display_name and
service_name. They were either always identical, or service_name was a
[slug](https://sentry.io/answers/slug-in-django/)-ified version of the
display_name.

Having two names is confusing, especially for https://github.com/ai-dynamo/dynamo/issues/2267 .

Now we only have `display_name`. The HTTP server gets to decide how to
translate that to the `model` HTTP field (if at all), it's not in the
card. In practice it still uses a slug.

Most of the code change in here is because I merged
`model_card/create.rs` and `model_card/model.rs`. It always confused me
why the object and it's implementation should be in separate files
(modules). It confused Rust too, we had to make fields `pub` just to access them.
I suspect it was a pattern from a different language mis-applied to Rust. Is fixed.