Fix default EndpointId values, remove silly AI test asserts
The lovely part is declaring e.g. `let namespace;` without a value means
the compiler ensures it always gets a value, no branch misses it. And
each branch does the minimum number of heap allocations.

That last one includes a full stack trace in debug build, so does not
match. Checking for not is_err and yes is_ok is redundant.