feat: Make part of discovery re-usable
Provide this function to fetch a ModelDeploymentCard from any part of
the code, without knowing how it's stored. Previously this was buried in
the discovery code.

```
ModelDeploymentCard::load_from_store(model_slug: &Slug, drt: &DistributedRuntime) -> anyhow::Result<Option<Self>>;
```

That will use whatever storage is configured on the runtime (currently
etcd and nats) to fetch the card.

Once that call returns the card is fully ready to use locally.

Signed-off-by: Graham King <grahamk@nvidia.com>