fix: Make openai_harmony setup call async and blocking
The `openai_harmony` crate calls `reqwest` in blocking mode. That means
it starts a tokio runtime inside. This is a problem because we are
already inside a runtime. We noticed this on shutdown in tests, when the internal
runtime tries to Drop, it panics because we're in an async context.

PR changes the `openai_harmony` setup to be in a blocking thread, so not
an async context. To do that ironically we need to make the setup async.

Unfortunately due to how async works, that means every caller all the
way up has to change to also being `async`. Thankfully `generate` is
already async, so it stopped at that point.

Signed-off-by: Graham King <grahamk@nvidia.com>