fix: If we detect model corruption log a warning
If we are already serving a model and a new worker arrives with a
different model card, claiming to be for the same model, we now don't
know which is the correct model. Log an error.

Ideally we would stop using that instance, but that happens in a
different crate. It's not clear how we can communicate with it, see
TODO.

Signed-off-by: Graham King <grahamk@nvidia.com>