chore: Remove GGUF support
- It doesn't work right now, it's a maintenance burden.
- I don't think anyone knows we have it.
- Customers are not interested.

This was an experiment in the early days of Dynamo when we were not
sure which direction it would go. All the interest is large data-center
scale deployments which are not running GGUF files.

Local dev is smooth enough with sglang and a small safetensors model (e.g. Qwen 0.6),
we don't need a GGUF file for that. Even if we did, the code paths were
different enough it wouldn't be a great test.

Less to support means we can focus better on what we have.

Signed-off-by: Graham King <grahamk@nvidia.com>