chore: Remove llama.cpp engine
Both the Rust / C++ one in `lib/engines/` and the Python / C++ one in
`components/src/dynamo/llama.cpp`.

llama.cpp is focused on single-node usage, typically for running
on personal machines. Dynamo is focused on data-center deployments,
typically multi-node. It didn't fit. Removing it allows us to focus
better.

Further, because llama.cpp wants to run GGUF files we had a whole
separate path in the frontend, which is now also gone (#3488).

Signed-off-by: Graham King <grahamk@nvidia.com>