feat: Introduce storage_client in DistributedRuntime
This will gradually replace `etcd_client`. It is backed by etcd
normally, or memory when etcd is not available.

Also use it to list instances, as an example and first usage.

This is the same interface that ModelDeploymentCards have always used
for publishing and reading, so it is well tested.

Signed-off-by: Graham King <grahamk@nvidia.com>