chore: Remove model_config from LocalModel
To support running GGUF files we allowed passing the path to an HF
repo. The weights would come from the GGUF file, but the model config
and tokenizer would come from the HF repo. That was because we didn't
support reading all the relevant fields from GGUF files yet.

GGUF files are gone in https://github.com/ai-dynamo/dynamo/pull/3488 ,
so there's no need for this feature.

Signed-off-by: Graham King <grahamk@nvidia.com>