feat(frontend): Get model config files (`tokenizer.json` et al.) from MX
Previously the backend would push all model config files into the NATS
object store, to make them available to the frontends. It then changed
the ModelDeploymentCard to have the nats URL instead of local path.

Now we do this instead:

Leave the local paths untouched in the model card in etcd. When frontend
fetches the model card it first checks if the same path is visible for
it too. This is good for local dev, but more importantly the K8s PVC is
often mounted on the same path on every pod, so likely the frontend can
already see the files.

If the files are not visible locally it downloads them from modelexpress, and
changes the local files to point to the new cache directory.

This makes publishing a model faster, because uploading / downloading the ~10 MiB
tokenizer took a few seconds. It also reduces our reliance on NATS.

The downside is now we can only use models that modelexpress can see.
Fix incoming for that.

Signed-off-by: Graham King <grahamk@nvidia.com>