refactor: Make `nats_client` optional internally
Dynamo will still fail to start if NATS is not available, we still need
it.

All internal uses of NATS are `Option<nats::Client>`, so that all
our code handles it not existing. Currently it either doesn't enable
that feature (metrics stats scraping), or returns an error (request
routing which is NATS only).

This will make it easier to run without NATS in the future.

Signed-off-by: Graham King <grahamk@nvidia.com>