chore: Remove static mode
Static mode was a way to run without `etcd`. It was relatively complex
to use and (as you can see) involved a lot of changes throughout the code.

We now have `storage/key_value_store/file.rs::FileStore` which mimics
etcd but from the local filesystem. It allows running without etcd with
far fewer code changes.

Signed-off-by: Graham King <grahamk@nvidia.com>