feat(key_value_store): FileStore entries have a lease that auto-expires
To mirror the `etcd` lease behavior, we start a background thread that
removes files older than 10s (same as etcd lease expiry). We `touch`
those files to update them more often than 10s, which keeps them alive.

That ensures we clean up keys from processes that did not shutdown
cleanly.

Once we get into the hundreds of machines with a shared file store this
would be too much disk activity. `FileStore` is not designed for that. The
main production deployment target is still etcd and Kubernetes.

Signed-off-by: Graham King <grahamk@nvidia.com>