nixl_connect: Change the way `bytes` are passed to NIXL
This change changes the way `bytes` are handled as data for the `Descriptor` class.

Instead of relying on `id()` which returns the memory address of the Python object itself,
this change adds a new method `_bytes_to_data_tuple` which returns a `tuple[int, int]` containing the memory address of the underlying data and its size.

The new method relies on ctypes to create an alias of the Python object and provide the actual address of the underlying data.
No data copies occur; only static type casting is used on the pointers to manipulate the interfaces presented by the ctypes library.

Despite the new method supporting both the `bytes` and `bytearray` types,
this change does not affect the types accepted by the `Descriptor` type's constructor.

Signed-off-by: J Wyman <jwyman@nvidia.com>