fix(runtime): Handle dropping the tokio runtime in async context
When used from the Python bindings the runtime will be dropped from (I think) Python's asyncio.
Tokio does not allow this and will panic. That panic prevents logging from printing it's last
messages, which makes knowing what went wrong very difficult.

This is the panic:
> pyo3_runtime.PanicException: Cannot drop a runtime in a context where blocking is not allowed.
> This happens when a runtime is dropped from within an asynchronous context.

Hence we wrap the runtime in a ManuallyDrop and use tokio's alternative shutdown if we detect
that we are inside an async runtime.

Signed-off-by: Graham King <grahamk@nvidia.com>