feat: nixl_connect: Improve Concurrency Support
This change moves the nixl_connect library from a persistent connector based design
to a per connection based design. When an active or passive operation is created,
a connection object is created to represent the local side of a connection to a remote worker.
The connection is responsible for keeping the descriptor data and operation state separated
when multiple operations are executing at the same time.

Prior to this change, it was possible for two operations to intersect leading to errors and
disconnections.

Signed-off-by: J Wyman <jwyman@nvidia.com>