fix(storage): File store shouldn't notify on metdata changes
File store starts an `inotify` style watch on the key-value store folder to
notice new models being registered. We also start a keep-alive thread
that touches the files every file seconds. That touch was triggering a
notification, so we thought there was a new model. It was harmless, we
detect the duplicate later, but it makes for noisy logs.

Caused by the intersection of #4250 and #4301.

Also clean up very verbose debug logs, and somehow the bindings Cargo.lock
needed updating.

Signed-off-by: Graham King <grahamk@nvidia.com>