fix(storage): Correctly encoding FileStore keys
Finishes the introduction of a new-type `Key` to represent a key-value
store Key. It is no longer Slug-ified. The storage backend will encode
depending on the characters it supports.

FileStore transparently percent-encodes / decodes it to make it a safe
filename. This removes the awkward and fragile replacement of slash with
underscore that FileStore was doing.

Signed-off-by: Graham King <grahamk@nvidia.com>