fix(dynamo-run): Run without etcd/nats, HTTP port to 8000
`dynamo-run` in all-in-one-mode now works without etcd and nats again,
as it did before the etcd/nats removal project.

all-in-one means not being a distributed system, the frontend and backend
run in the same process:
```
dynamo-run in=http out=mistralrs ...
```

Also change the HTTP port to 8000 to match Python frontend.

Signed-off-by: Graham King <grahamk@nvidia.com>