chore(pipeline): Move migration outside of backend
`Backend` is going to move to, well, the backend (worker). Migration
now works from `BackendOutput` not `LLMEngineOutput`.

The router also needs to change, but PrefillRouter (`prefill_op`) complicates matters.
We need to change it and the router at the same time, once the worker
returns `BackendOutput` over the wire instead of `LLMEngineOutput`.

Signed-off-by: Graham King <grahamk@nvidia.com>