chore: De-tokenize in the worker
Add a `Backend` layer on the worker side. This handles de-tokenization.

Note that this only adds it for the python engines. The Rust engines
(mocker and dynamo-run) already did this.

`register_llm` now saves the `ModelDeploymentCard` (MDC) that we registered in
the bindings `Endpoint`. It is retrieved by `serve_endpoint` and used to
get the tokenizer to build the pipeline. Ideally the MDC would be on the
main Rust `Endpoint`, but that is in lib/runtime so it doesn't know
about LLMs.

This PRs leaves `Backend` in the frontend. It already detects if the
request is de-tokenized and short-circuits, so this doesn't slow
us down. In a later PR we can remove that. It involves changes to the router,

Also removed some noisy Discovery debug statements.

Signed-off-by: Graham King <grahamk@nvidia.com>