feat: Pass tool definitions to parsers
* Why?

The qwen3 coder parser in vLLM needs access to the tool's type
declarations in order to parse things as expected. Without this,
certain tool calling benchmarks (such as BFCL-3) will fail on samples
where the parsed types matter.

* What?

This commit extends the tool call parsing machinery to extract such
information from the request and pipes all the way down to the parsers.